requirements.txt
streamlit
python-dotenv
huggingface_hub
# streamlit firebase-admin pyrebase4
firebase